Non Decaying Agent, Convergence
Added a non-decaying epsilon greedy agent, which maintains an epsilon
value of 0.05 throughout. Added convergence of arm belief to both types
of agents. This uses samples of the posterior distributions for the
Bayesian agent (i.e. an "elegant" approach) compared to the decaying
epsilon which consistently has the same iteration of convergence.